requirements.txt
dj-database-url==0.5.0
flake8>=3.0
sqids>=0.3
mock==5.1.0
psycopg2-binary==2.9.3
pytest-cov>=4.1.0
pytest-django==4.5.2
pytest-sugar==0.9.7
pytest>=7.4.2
Sphinx>=7.2.6
tox-pyenv==1.1.0
tox==4.11.3
twine==4.0.2
Django>=3.2,<4.3
djangorestframework>=3.11.2,<3.15
orjson==3.9.7
black==23.9.1
isort==5.12.0
pre-commit==3.4.0
pylint==2.17.5
